Add context lines to SBLGNT Apparatus
These lines represent words directly before or after a variation group
that have the same surface text as one of the words inside the variation
group. Logos' Word Numbers sometimes "swap" these context words for the
inside word if the rest of the group is omitted.

For example, "A and B and C" can be marked up in the apparatus so that
"and B" is omitted in an edition, while in the word numbers, "B and"
(the and from the context) is omitted.